@@ -8,7 +8,9 @@
     "build": "craco build",
     "test": "craco test",
     "eject": "craco eject",
-    "prettier": "prettier --write ."
+    "prettier": "prettier --write .",
+    "test:ply": "npx playwright test",
+    "test:ply-ui": "npx playwright test --ui"
   },
   "dependencies": {
     "@emotion/react": "^11.11.1",
@@ -39,12 +41,15 @@
     "@babel/plugin-proposal-private-property-in-object": "^7.21.11",
     "@babel/plugin-transform-private-property-in-object": "^7.22.11",
     "@craco/craco": "^7.1.0",
+    "@playwright/test": "^1.41.2",
+    "@types/node": "^20.11.17",
     "@types/react": "^17.0.0",
+    "@types/react-dom": "^18.2.19",
     "prettier": "^3.2.5",
     "react-google-recaptcha": "^3.1.0",
     "react-google-recaptcha-enterprise": "^1.0.3",
     "tailwindcss": "^3.3.5",
-    "typescript": "^3.2.1"
+    "typescript": "^4.0.0"
   },
   "eslintConfig": {
     "extends": [
